requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
pyhive>=0.6.0
trino>=0.321.0
presto-python-client>=0.8.0

# AI/ML dependencies
openai>=1.0.0
anthropic>=0.7.0
langchain>=0.1.0
langchain-openai>=0.0.5
scikit-learn>=1.3.0
transformers>=4.30.0

# Data analysis
great-expectations>=0.17.0
ydata-profiling>=4.0.0

# Web framework
fastapi>=0.100.0
uvicorn>=0.22.0
streamlit>=1.25.0

# Database connections
psycopg2-binary>=2.9.0
pymongo>=4.0.0
redis>=4.5.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
typer>=0.9.0
rich>=13.0.0
loguru>=0.7.0
pyyaml>=6.0.0
jinja2>=3.1.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0
